% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BR_get_schedule.R
\name{BR_get_schedule}
\alias{BR_get_schedule}
\title{Retrieves season schedule}
\usage{
BR_get_schedule(Year, Month)
}
\arguments{
\item{Year}{year in YYYY format}

\item{Month}{full name of month, all in lowercase}
}
\value{
a data frame with box scores for specified month
}
\description{
`BR_get_schedule` gets the full season schedule from basketball-reference
}
\examples{
BR_get_schedule(Year = 2019, Month = 'october')
}
