% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roster.R
\name{get_roster}
\alias{get_roster}
\title{retrieves team roster}
\usage{
get_roster(TeamID)
}
\arguments{
\item{TeamID}{NBA id of team to get data for}
}
\value{
a data frame with roster data for a given team
}
\description{
`synergy_team` gets roster data for a given team
}
\examples{
get_roster(TeamID = 1610612745)
}
