% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy_player.R
\name{synergy_player}
\alias{synergy_player}
\title{retrieves synergy player stats}
\usage{
synergy_player(Category, Year = 2018, Season.seg = "Reg",
  Off.def = "offensive", Player.limit = 500)
}
\arguments{
\item{Category}{category of stat to get}

\item{Year}{start year of season schedule in YYYY format}

\item{Season.seg}{part of season}

\item{Off.def}{choose to retrieve offensive or defensive data}

\item{Player.limit}{limit of players to get data for}
}
\value{
a data frame with team synergy data of a given category
}
\description{
`synergy_player` gets player synergy data for a given category
}
\examples{
synergy_player(Category = 'transition', Off.def = 'defensive')
}
