% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_performance.R
\name{team_performance}
\alias{team_performance}
\title{Retrieves team season performance data}
\usage{
team_performance(TeamID, Year = 2018)
}
\arguments{
\item{TeamID}{NBA id of team}

\item{Year}{start year of season schedule in YYYY format}
}
\value{
a data frame with performance data for a given team for a given year
}
\description{
`team_performance` gets performance data for a full season
}
\examples{
team_performance(TeamID = 1610612745, Year = 2018)
}
