% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_list.R
\name{teams_list}
\alias{teams_list}
\title{Retrieves list of teams}
\usage{
teams_list(Year, Type = "standard")
}
\arguments{
\item{Year}{start year of season schedule in YYYY format}

\item{Type}{type of data to keep (standard, vegas)}
}
\value{
a data frame with basic franchise information for all NBA teams
}
\description{
`teams_list` gets basic information on all NBA teams
}
\examples{
teams_list(Year = 2018, Type = 'vegas')
}
